Fix filtered selection in FcCardTableRequests
To simplify this, I've also moved the implementation of `FcCardTableRequests` into `FcRequestsTrack` - this avoids a lot of multi-level passing around of state and `$emit` hackery / overhead.
